refactor: MovieConverter의 convertFromMovieToResponseFindCurrentMovie 메소드 사용하도록 변경
