+++ src/main/java/prgrms/marco/be02marbox/domain/theater/service/ScheduleService.java
@@ -8,11 +8,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import prgrms.marco.be02marbox.domain.movie.Movie;
+import prgrms.marco.be02marbox.domain.movie.dto.ResponseFindCurrentMovie;
 import prgrms.marco.be02marbox.domain.movie.repository.MovieRepository;
+import prgrms.marco.be02marbox.domain.movie.service.utils.MovieConverter;
 import prgrms.marco.be02marbox.domain.theater.Schedule;
 import prgrms.marco.be02marbox.domain.theater.TheaterRoom;
 import prgrms.marco.be02marbox.domain.theater.dto.RequestCreateSchedule;
-import prgrms.marco.be02marbox.domain.theater.dto.ResponseFindCurrentMovie;
 import prgrms.marco.be02marbox.domain.theater.repository.ScheduleRepository;
 import prgrms.marco.be02marbox.domain.theater.repository.TheaterRoomRepository;
 import prgrms.marco.be02marbox.domain.theater.service.utils.ScheduleConverter;
@@ -23,18 +24,21 @@ public class ScheduleService {
 	private static final int CURRENT_SCHEDULE_PERIOD = 19;
 
 	private final ScheduleRepository scheduleRepository;
-	private final ScheduleConverter converter;
+	private final ScheduleConverter scheduleConverter;
 	private final TheaterRoomRepository theaterRoomRepository;
 	private final MovieRepository movieRepository;
+	private final MovieConverter movieConverter;
 
 	public ScheduleService(ScheduleRepository scheduleRepository,
 		ScheduleConverter converter,
 		TheaterRoomRepository theaterRoomRepository,
-		MovieRepository movieRepository) {
+		MovieRepository movieRepository,
+		MovieConverter movieConverter) {
 		this.scheduleRepository = scheduleRepository;
-		this.converter = converter;
+		this.scheduleConverter = converter;
 		this.theaterRoomRepository = theaterRoomRepository;
 		this.movieRepository = movieRepository;
+		this.movieConverter = movieConverter;
 	}
 
 	@Transactional
@@ -44,7 +48,8 @@ public class ScheduleService {
 		Movie movie = movieRepository.findById(requestCreateSchedule.movieId())
 			.orElseThrow(() -> new IllegalArgumentException("존재하지 않는 영화 ID"));
 
-		Schedule schedule = converter.convertFromRequestCreateScheduleToSchedule(requestCreateSchedule, theaterRoom,
+		Schedule schedule = scheduleConverter.convertFromRequestCreateScheduleToSchedule(requestCreateSchedule,
+			theaterRoom,
 			movie);
 
 		Schedule savedSchedule = scheduleRepository.save(schedule);
@@ -59,7 +64,7 @@ public class ScheduleService {
 			LocalDate.now().plusDays(CURRENT_SCHEDULE_PERIOD));
 
 		return scheduleList.stream()
-			.map(converter::convertFromScheduleToResponseFindCurrentMovie)
+			.map(schedule -> movieConverter.convertFromMovieToResponseFindCurrentMovie(schedule.getMovie()))
 			.distinct()
 			.collect(Collectors.toList());
 	}
